requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
newspaper3k==0.2.8
python-dotenv==1.0.0